Make stdts a top level namespace
Summary:
stdts to be moved into another repo someday as a micro library.

I don't want to `using namespace scraps` anywhere outside of scraps. Same with all our libraries. The only acceptable `using namespace ...` is the namespace of the library in which you're working. This assists in that goal.

Test Plan: n/a

Reviewers: vmrob

Reviewed By: vmrob

Subscribers: #engineering

Differential Revision: https://phabricator.btlive.tv/D1623