reduce run loop writing
Summary: reduce run loop writing. the error we see on the ingest server is the "try again" error, which means we're trying to write to a pipe whose buffer is already full. not writing excessively will fix that warning (which was in fact harmless)

Test Plan: unit. nothing should change

Reviewers: vmrob

Reviewed By: vmrob

Subscribers: #engineering

Differential Revision: https://phabricator.btlive.tv/D1808