Remove OpenSSL build concurrency in CI on MacOS
Summary:
Due to OpenSSL's lack of proper build concurrency and overuse of recursive makefiles, hundreds of processes were spawned simultaneously when building concurrently. Normally this is a non-issue, but for MacOS, the OpenSSL builds sometimes fail in CI--presumably due to user limits set on those processes by the CI system.

This patch prevents OpenSSL from building concurrently in the event that the CI environment variable is defined.

Test Plan: manual

Reviewers: cbrown

Reviewed By: cbrown

Subscribers: #engineering

Differential Revision: https://phabricator.btlive.tv/D1991