Add PollingFutureSynchronizer
Summary: The same function as FutureSynchronizer, but it polls the futures when `update()` is called, and calls the callbacks synchronously. There are potential use cases for both. The use of FutureSynchronizer in MainView wasn't safe during the removal of MainView from the view hierarchy during shutdown.

Test Plan: Unit tests on their way

Reviewers: mike.naquin

Reviewed By: mike.naquin

Subscribers: #engineering

Differential Revision: https://phabricator.btlive.tv/D2046